+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # 0.0.2
 - Bugfix OTPScreen with Icon Animation Handler
+- Refactor otp to viewmodel
 
 # 0.0.1
 - Create SplashScreen